refactor: Moved identifier lookup logic to service
Moved lookup and create of identifiers/namespaces to IdentifierService, with BaseTIRS able to catch and rethrow any understood exceptions as TIRSExceptions instead
